requirements.txt
"""
Flask
requests
python-dotenv
firebase-admin
"""